Fixed problems processing MulitPath Echo images
Seimann's software now (2/1/2024) produces MulitPath Echo images.  The flywheel representation
for these images are not combatible with nibabel.

It seems that nibabel was only being used to get the number of rows and columns in the images.  Removed
dependancy on nibabel and get the row and column numbers directly from the flywheel image metadata.

-Gaylord